ct_dashboard_api: extract password business logic to api layer
