+++ backend/ct_dashboard_api/src/routes/v1/reset_password_confirm.ts
@@ -1,23 +1,14 @@
-import { hashPassword } from "@oko-wallet/crypto-js";
+import { ErrorCodeMap } from "@oko-wallet/oko-api-error-codes";
 import { registry } from "@oko-wallet/oko-api-openapi";
 import { ErrorResponseSchema } from "@oko-wallet/oko-api-openapi/common";
 import {
   ResetPasswordConfirmRequestSchema,
   ResetPasswordConfirmSuccessResponseSchema,
 } from "@oko-wallet/oko-api-openapi/ct_dashboard";
-import {
-  getCTDUserWithCustomerAndPasswordHashByEmail,
-  updateCustomerDashboardUserPassword,
-} from "@oko-wallet/oko-pg-interface/customer_dashboard_users";
-import { verifyEmailCode } from "@oko-wallet/oko-pg-interface/email_verifications";
 import type { OkoApiResponse } from "@oko-wallet/oko-types/api_response";
 import type { Request, Response } from "express";
 
-import {
-  CHANGED_PASSWORD_MAX_LENGTH,
-  CHANGED_PASSWORD_MIN_LENGTH,
-  PASSWORD_CONTAINS_NUMBER_REGEX,
-} from "@oko-wallet-ctd-api/constants";
+import { resetPasswordConfirmRequest } from "@oko-wallet-ctd-api/api/customer_password";
 
 registry.registerPath({
   method: "post",
@@ -76,97 +67,14 @@ export async function resetPasswordConfirm(
   req: Request,
   res: Response<OkoApiResponse<{ message: string }>>,
 ) {
-  try {
-    const state = req.app.locals;
-    const { email, code, newPassword } = req.body;
+  const state = req.app.locals;
 
-    if (!email || !code || !newPassword) {
-      res.status(400).json({
-        success: false,
-        code: "INVALID_REQUEST",
-        msg: "Missing fields",
-      });
-      return;
-    }
+  const result = await resetPasswordConfirmRequest(state.db, req.body);
 
-    if (newPassword.length < CHANGED_PASSWORD_MIN_LENGTH) {
-      res.status(400).json({
-        success: false,
-        code: "INVALID_EMAIL_OR_PASSWORD",
-        msg: `Password must be at least ${CHANGED_PASSWORD_MIN_LENGTH} characters long`,
-      });
-      return;
-    }
-
-    if (newPassword.length > CHANGED_PASSWORD_MAX_LENGTH) {
-      res.status(400).json({
-        success: false,
-        code: "INVALID_EMAIL_OR_PASSWORD",
-        msg: `Password must be at most ${CHANGED_PASSWORD_MAX_LENGTH} characters long`,
-      });
-      return;
-    }
-
-    if (!PASSWORD_CONTAINS_NUMBER_REGEX.test(newPassword)) {
-      res.status(400).json({
-        success: false,
-        code: "INVALID_EMAIL_OR_PASSWORD",
-        msg: "Password must include at least one number",
-      });
-      return;
-    }
-
-    const verificationResult = await verifyEmailCode(state.db, {
-      email,
-      verification_code: code,
-    });
-
-    if (!verificationResult.success) {
-      res.status(400).json({
-        success: false,
-        code: "INVALID_VERIFICATION_CODE",
-        msg: "Invalid or expired verification code",
-      });
-      return;
-    }
-
-    const customerAccountResult =
-      await getCTDUserWithCustomerAndPasswordHashByEmail(state.db, email);
-
-    if (!customerAccountResult.success || !customerAccountResult.data) {
-      res.status(404).json({
-        success: false,
-        code: "CUSTOMER_ACCOUNT_NOT_FOUND",
-        msg: "User not found",
-      });
-      return;
-    }
-
-    const hashedNewPassword = await hashPassword(newPassword);
-    const updateResult = await updateCustomerDashboardUserPassword(state.db, {
-      user_id: customerAccountResult.data.user.user_id,
-      password_hash: hashedNewPassword,
-    });
-
-    if (!updateResult.success) {
-      res.status(500).json({
-        success: false,
-        code: "FAILED_TO_UPDATE_PASSWORD",
-        msg: "Failed to update password",
-      });
-      return;
-    }
-
-    res.status(200).json({
-      success: true,
-      data: { message: "Password reset successfully" },
-    });
-  } catch (error) {
-    console.error("Reset password confirm error:", error);
-    res.status(500).json({
-      success: false,
-      code: "UNKNOWN_ERROR",
-      msg: "Internal server error",
-    });
+  if (!result.success) {
+    res.status(ErrorCodeMap[result.code] ?? 500).json(result);
+    return;
   }
+
+  res.status(200).json(result);
 }
